# Lab9/backend/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get list of courses",
                "description": "Returns list of courses",
                "operationId": "getCourseList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a new course",
                "description": "Create a new course record",
                "operationId": "createCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/courses/{course}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get a course by ID",
                "description": "Returns a single course",
                "operationId": "getCourseById",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "$ref": "#/components/schemas/Course"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update an existing course",
                "description": "Update an existing course record",
                "operationId": "updateCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Course"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "description": "Deletes a single course",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "/api/courses/{course}/edit": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Show course edit form",
                "description": "Display the form for editing an existing course",
                "operationId": "editCourseForm",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "ID of the course to edit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Course": {
                "title": "Course",
                "description": "Course model schema",
                "required": [
                    "id",
                    "course_level_id",
                    "name",
                    "credit_theory",
                    "credit_lab"
                ],
                "properties": {
                    "id": {
                        "description": "Course ID",
                        "type": "integer",
                        "example": 1
                    },
                    "course_level_id": {
                        "description": "ID of the course level",
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "description": "Name of the course",
                        "type": "string",
                        "example": "Introduction to Programming"
                    },
                    "name_vn": {
                        "description": "Name of the course in Vietnamese",
                        "type": "string",
                        "example": "Giới thiệu về lập trình"
                    },
                    "credit_theory": {
                        "description": "Number of theoretical credits",
                        "type": "integer",
                        "example": 3
                    },
                    "credit_lab": {
                        "description": "Number of laboratory credits",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "description": "Description of the course",
                        "type": "string",
                        "example": "This course covers basic concepts of programming including algorithms, data structures, and software engineering principles."
                    }
                },
                "type": "object"
            },
...
        }
    }
}